+++ src/Reflex/Vty/Host.hs
@@ -222,7 +222,8 @@ runVtyApp app = do
   vty <- getDefaultVty
   runVtyAppWithHandle vty app
 
--- | Returns the standard vty configuration with mouse and focus tracking enabled.
+-- | Returns the standard vty configuration with mouse, focus tracking,
+-- and bracketed paste enabled.
 getDefaultVty :: IO V.Vty
 getDefaultVty = do
   cfg <- V.userConfig
@@ -230,4 +231,5 @@ getDefaultVty = do
   liftIO $ do
     V.setMode (V.outputIface vty) V.Mouse True
     V.setMode (V.outputIface vty) V.Focus True
+    V.setMode (V.outputIface vty) V.BracketedPaste True
   return vty